+++ ezdxf/entities/blockrecord.py
@@ -102,7 +102,7 @@ class BlockRecord(DXFEntity):
         dxf = super().load_dxf_attribs(processor)
         if processor:
             tags = processor.load_dxfattribs_into_namespace(dxf, acdb_blockrec)
-            if len(tags):  # not supported by DXF R12
+            if len(tags) and False:  # deliberately disabled
                 processor.log_unprocessed_tags(tags, subclass=acdb_blockrec.name)
         return dxf